delete-room

+++ admin/src/components/datatable/Datatable.jsx
@@ -7,20 +7,30 @@ import useFetch from '../../hooks/useFetch';
 import { userApi } from '../../api/apiConfig';
 import axios from 'axios';
 
-const Datatable = ({ columns }) => {
+const Datatable = ({ columns, action = true, pageSize = 9 }) => {
   const location = useLocation();
-  const path = location.pathname.split('/')[1];
+  const pathName = location.pathname.split('/')[1];
+  const path = pathName === '' ? 'transactions?limit=8' : pathName;
   const [list, setList] = useState();
+  const [id, setId] = useState(null);
   const { data, loading, error } = useFetch(`/${path}`);
   useEffect(() => {
     setList(data);
   }, [data]);
 
   const handleDelete = async id => {
-    try {
-      await axios.delete(`/${path}/${id}`);
-      setList(list.filter(item => item._id !== id));
-    } catch (error) {}
+    if (window.confirm('Are your sure?')) {
+      try {
+        const result = await axios.delete(`/${path}/${id}`);
+        if (result.status === 200) {
+          alert(result.data);
+          setList(list.filter(item => item._id !== id));
+        }
+      } catch (error) {
+        alert(error?.response.data.message);
+        console.log('error :>> ', error);
+      }
+    }
   };
 
   const actionColumn = [
@@ -46,25 +56,34 @@ const Datatable = ({ columns }) => {
     },
   ];
   return (
-    <div className='datatable'>
-      <div className='datatableTitle'>
-        Add New
-        <Link to={`/${path}/new`} className='link'>
-          Add New
-        </Link>
-      </div>
-      {list && (
-        <DataGrid
-          className='datagrid'
-          rows={list}
-          columns={columns.concat(actionColumn)}
-          pageSize={9}
-          rowsPerPageOptions={[9]}
-          checkboxSelection
-          getRowId={row => row._id}
-        />
+    <>
+      {loading ? (
+        'loading...'
+      ) : (
+        <div className='datatable'>
+          {action && (
+            <div className='datatableTitle'>
+              Add New
+              <Link to={`/${path}/new`} className='link'>
+                Add New
+              </Link>
+            </div>
+          )}
+          {list && (
+            <DataGrid
+              className='datagrid'
+              rows={list}
+              // columns={columns}
+              columns={action ? columns.concat(actionColumn) : columns}
+              pageSize={pageSize}
+              rowsPerPageOptions={[9]}
+              checkboxSelection
+              getRowId={row => row._id}
+            />
+          )}
+        </div>
       )}
-    </div>
+    </>
   );
 };
