+++ backend/controllers/room.js
@@ -1,5 +1,7 @@
 const Room = require('../models/Room');
 const Hotel = require('../models/Hotel');
+const Transaction = require('../models/Transaction');
+const createHttpError = require('http-errors');
 
 exports.createRoom = async (req, res, next) => {
   const hotelId = req.params.hotelid;
@@ -48,22 +50,39 @@ exports.updateRoomAvailability = async (req, res, next) => {
   }
 };
 exports.deleteRoom = async (req, res, next) => {
-  const hotelId = req.params.hotelid;
-  const roomId=req.params.roomid
+  const roomId = req.params.id;
   try {
+    const transaction = await Transaction.findOne({ 'rooms.roomId': roomId });
+    if (transaction)
+      return next(
+        createHttpError(404, 'The room with transactions cannot be deleted.')
+      );
+
     await Room.findByIdAndDelete(roomId);
-    try {
-      await Hotel.findByIdAndUpdate(hotelId, {
-        $pull: { rooms: roomId },
-      });
-    } catch (err) {
-      next(err);
-    }
-    res.status(200).json('Room has been deleted.');
-  } catch (err) {
-    next(err);
+    await Hotel.updateMany({ rooms: roomId }, { $pull: { rooms: roomId } });
+
+    res.status(200).json(`Room has been deleted.`);
+  } catch (error) {
+    next(error);
   }
 };
+// exports.deleteRoom = async (req, res, next) => {
+//   const hotelId = req.params.hotelid;
+//   const roomId=req.params.roomid
+//   try {
+//     await Room.findByIdAndDelete(roomId);
+//     try {
+//       await Hotel.findByIdAndUpdate(hotelId, {
+//         $pull: { rooms: roomId },
+//       });
+//     } catch (err) {
+//       next(err);
+//     }
+//     res.status(200).json('Room has been deleted.');
+//   } catch (err) {
+//     next(err);
+//   }
+// };
 // user
 exports.getRoom = async (req, res, next) => {
   try {
